new features: moved web3 connection logic to a hook and added logic to the components
