+++ src/components/App/App.js
@@ -6,13 +6,14 @@ import FirstStepPage from '../../pages/step1-page/step1-page';
 import ProofPage from '../../pages/proof-page/proof-page';
 import DelegationPage from '../../pages/delegation-page/delegation-page';
 import DonePage from '../../pages/done-page/done-page';
-
 import ConnectWallet from '../ConnectWallet/ConnectWallet';
-
+import { useWeb3Connection } from '../../hooks/useWeb3Connection';
 
 import './App.css';
 
 function App() {
+  const { address } = useWeb3Connection()
+
   return (
     <div className="App">
       <Routes>
