+++ src/components/ConnectWallet/ConnectWallet.js
@@ -1,125 +1,19 @@
-import WalletConnectProvider from '@walletconnect/web3-provider'
-import { providers } from 'ethers'
-import { useCallback, useEffect, useReducer } from 'react'
-import { useDispatch, useSelector } from 'react-redux'
-import Web3Modal from 'web3modal'
-import { resetWeb3Provider, setAddress, setWeb3Provider } from '../../store/actions/wallet'
-import { ellipseAddress, getChainData } from '../../utils'
-
-const { REACT_APP_INFURA_KEY: INFURA_ID } = process.env
-
-const providerOptions = {
-    walletconnect: {
-        package: WalletConnectProvider, // required
-        options: {
-            infuraId: INFURA_ID, // required
-        }
-    }
-}
-
-let web3Modal
-if  (typeof window !== 'undefined') {
-        web3Modal = new Web3Modal({
-            network: 'mainnet', // optional
-            cacheProvider: true,
-            providerOptions, // required
-    })
-}
+import Button from '../Button/Button'
+import { useWeb3Connection } from '../../hooks/useWeb3Connection'
+import { getChainData } from '../../utils'
 
 const ConnectWallet = () => {
-    const state = useSelector(state => state.wallet)
-    const dispatch = useDispatch()
-    const { provider, web3Provider, address, chainId } = state
+    const { connect, disconnect, chainId, web3Provider } = useWeb3Connection()
+    const chainData = getChainData(chainId)
 
-    const connect = useCallback(async () => {
-        const provider = await web3Modal.connect()
-        const web3Provider = new providers.Web3Provider(provider)
-        const signer = web3Provider.getSigner()
-        const address = await signer.getAddress()
-        const network = await web3Provider.getNetwork()
-
-        dispatch(setWeb3Provider(provider, web3Provider, address, network))
-    }, [])
-
-    const disconnect = useCallback(
-        async function () {
-            await web3Modal.clearCachedProvider()
-            if (provider?.disconnect && typeof provider.disconnect === 'function') {
-                await provider.disconnect()
-            }
-            dispatch(resetWeb3Provider())
-        },
-        [provider]
-    )
-
-    useEffect(() => {
-        if (web3Modal.cachedProvider) {
-            connect()
-        }
-    }, [connect])
-
-    useEffect(() => {
-        if (provider?.on) {
-        const handleAccountsChanged = (accounts) => {
-            // console.log('accountsChanged', accounts)
-            dispatch(setAddress(accounts))
-        }
-        const handleChainChanged = (_hexChainId) => {
-            window.location.reload()
-        }
-
-        const handleDisconnect = (error) => {
-            console.log('disconnect', error)
-            disconnect()
-        }
-
-        provider.on('accountsChanged', handleAccountsChanged)
-        provider.on('chainChanged', handleChainChanged)
-        provider.on('disconnect', handleDisconnect)
-
-        // Subscription Cleanup
-        return () => {
-            if (provider.removeListener) {
-                provider.removeListener('accountsChanged', handleAccountsChanged)
-                provider.removeListener('chainChanged', handleChainChanged)
-                provider.removeListener('disconnect', handleDisconnect)
-            }
-        }
-    }
-  }, [provider, disconnect])
-
-  const chainData = getChainData(chainId)
-
-  return (
-    <>
-        <div>
-            {address && (
-                <div className="grid">
-                    <div>
-                        <p className="mb-1">Network:</p>
-                        <p>{chainData?.name}</p>
-                    </div>
-                    <div>
-                        <p className="mb-1">Address:</p>
-                        <p>{ellipseAddress(address)}</p>
-                    </div>
-                </div>
-            )}
-        </div>
-
-        <main>
+    return (
+        <>
             {web3Provider ? (
-                <button className="button" type="button" onClick={disconnect}>
-                    Disconnect
-                </button>
+                <Button callback={disconnect} text='Disconnect'/>
             ) : (
-                <button className="button" type="button" onClick={connect}>
-                    Connect
-                </button>
+                <Button callback={connect} text='Connect a wallet'/>
             )}
-        </main>
-
-    </>
+        </>
   )
 }
